Migrate integrity.cy to playwright (#2901)
